+++ cargo-crev/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/dpc/crev/compare/cargo-crev-v0.5.0...HEAD)
+### Added
+
+- `cargo crev edit config` allows interactive user config edition
+- `open-cmd` in user config for customizing `cargo crev open` command
 
 ## [0.5.0](https://github.com/dpc/crev/compare/cargo-crev-v0.4.0...cargo-crev-v0.5.0) - 2019-03-06
 ### Added